shearing tests and function
